Created get_dg_rankings and prepped prefix in __connect_api
During retrieval of get-rankings from the API, it was noticed that there is a prefix in some of the API calls. This was included in
the __connect_api function call and utilized in the new get_deg_rankings function.

get_dg_rankings successfully retrieves datagolf.com's rankings for the upcoming week and parses the player's names depending on the
desired outcome.